Primeros pasos para agregar el Firebase
