+++ src/components/Register.js
@@ -22,22 +22,59 @@ export const Register = () => {
   // Se crea contenedor para inicio de sesión con correo, google y facebook
   const section1 = document.createElement('section');
   section1.className = 'section1';
+
+  const nameAndLastNameContainer = document.createElement('section');
+  nameAndLastNameContainer.className = 'name-lastName-container';
+
+  const inputName = document.createElement('input');
+  inputName.setAttribute('requiered', '');
+  inputName.setAttribute('placeholder', 'Nombre');
+  inputName.setAttribute('id', 'name');
+
+  const inputLastName = document.createElement('input');
+  inputLastName.setAttribute('requiered', '');
+  inputLastName.setAttribute('placeholder', 'Apellido');
+  inputLastName.setAttribute('id', 'lastName');
+
   const inputEmail = document.createElement('input');
   inputEmail.setAttribute('requiered', '');
   inputEmail.setAttribute('placeholder', 'Correo electrónico');
   inputEmail.setAttribute('id', 'emailLogin');
 
+  const inputPassword = document.createElement('input');
+  inputPassword.setAttribute('requiered', '');
+  inputPassword.setAttribute('placeholder', 'Contraseña');
+  inputPassword.setAttribute('id', 'password');
+
+  const inputConfirmPassword = document.createElement('input');
+  inputConfirmPassword.setAttribute('requiered', '');
+  inputConfirmPassword.setAttribute('placeholder', 'Confirma tu Contraseña');
+  inputConfirmPassword.setAttribute('id', 'confirm-password');
+
+  nameAndLastNameContainer.append(inputName, inputLastName);
+
   const buttonSignUp = document.createElement('button');
-  buttonSignUp.textContent = 'Registrarse';
+  buttonSignUp.textContent = 'Registrate!';
+
   const google = document.createElement('a');
   google.setAttribute('href', '');
   google.textContent = 'Registrarse con google';
   google.className = 'google';
+
   const facebook = document.createElement('a');
   facebook.setAttribute('href', '');
   facebook.textContent = 'Registrarse con facebook';
   facebook.className = 'facebook';
-  section1.append(inputEmail, buttonSignUp, google, facebook);
+  section1.append(
+    inputEmail,
+    nameAndLastNameContainer,
+    inputEmail,
+    inputPassword,
+    inputConfirmPassword,
+    buttonSignUp,
+    google,
+    facebook,
+  ); // SECTION 1 APPEND//
 
   // Se crea contendeor pra registrarse si no tien una cuenta
   const section2 = document.createElement('section');
@@ -51,10 +88,10 @@ export const Register = () => {
   section2.append(account, linkRegister);
 
   buttonSignUp.addEventListener('click', () => {
-    onNavigate('/register2');
+    onNavigate('/wall');
   });
   linkRegister.addEventListener('click', () => {
-    onNavigate('/wall');
+    onNavigate('/');
   });
 
   div.append(header, section1, section2);
