+++ src/components/Wall.js
@@ -2,28 +2,21 @@ import { onNavigate } from '../main.js';
 
 export const Wall = () => {
   // contenedor que almacenará los 2 botones y dará un solo return
-  const div = document.createElement('div');
+  const header = document.createElement('header');
+  header.className = 'header';
   // se crea titulo y parrafo
   const title = document.createElement('h1');
-  const paragraph = document.createElement('h2');
+  title.className = 'title';
   // Se crean 2 botones
-  const button = document.createElement('button');
-  const buttonBack = document.createElement('button');
-  const inputEmail = document.createElement('input');
-  const inputPass = document.createElement('input');
 
-  paragraph.textContent = 'Inicia Sesión';
-  button.textContent = 'Entrar';
-  buttonBack.textContent = 'Regresar';
+  const buttonBack = document.createElement('button');
+
   title.textContent = 'Yummy Friends';
-  button.addEventListener('click', () => {
-    onNavigate('/');
-  });
   buttonBack.addEventListener('click', () => {
     onNavigate('/');
   });
 
-  div.append(title, paragraph, inputEmail, inputPass, button, buttonBack);
+  header.append(title, buttonBack);
 
-  return div;
+  return header;
 };
