+++ src/index.html
@@ -14,6 +14,6 @@
 <body>
   <div id="root"></div> <!-- Nodo root: En el se va a renderizar toda la app-->
   <script type="module" src="main.js"></script>
+  <script type="module" src="./lib/firebase.js"></script>
 </body>
-
 </html>
